Gdbserver: move gdbserver commands to separate module.
